requirements.txt
python-dotenv
streamlit
pandas
pyarrow
pymupdf
chromadb
sentence-transformers
langchain
langchain-core
langchain-ollama
langchain-text-splitters
openpyxl
langchain-huggingface
langchain-chroma
pillow
pytesseract
ipykernel
pytest